# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend App Ecommerce API",
        "description": "Documentación de la API del Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3001",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/checkout/mis-pedidos": {
            "get": {
                "tags": [
                    "Checkout Web"
                ],
                "summary": "Listar los pedidos del cliente autenticado",
                "description": "Obtiene el historial de compras (pedidos) del usuario logueado.",
                "operationId": "8c18d62d0305db6cfe66980eff6c4f17",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos obtenida correctamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Checkout Web"
                ],
                "summary": "Procesar un nuevo pedido (Checkout)",
                "description": "Genera una nueva venta en estado PENDIENTE desde el carrito del cliente de la tienda virtual. No genera comprobante ni transacción financiera aún.",
                "operationId": "ea51cdcea2c3fda2327071ed2bb20819",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "direccion_envio_id",
                                    "tipo_metodo_pago_id",
                                    "costo_envio",
                                    "detalles"
                                ],
                                "properties": {
                                    "direccion_envio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_metodo_pago_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "costo_envio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "producto_id",
                                                "cantidad"
                                            ],
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cantidad": {
                                                    "type": "integer",
                                                    "example": 2,
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido registrado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación o stock insuficiente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}